Add Spell to server
