+++ server/library.ts
@@ -1,4 +1,5 @@
 import fs = require("fs");
+import _ = require("lodash");
 
 const sourceAbbreviations = {
     "monster-manual": "mm",
@@ -26,10 +27,29 @@ export interface Listing {
     Link: string;
 }
 
-export default class Library<TItem extends LibraryItem> {
+export interface StatBlock {
+    Name: string;
+    Id: string;
+    Type: string;
+    Source: string;
+}
+
+export const GetStatBlockKeywords = (statBlock: StatBlock) => statBlock.Type.replace(/[^\w\s]/g, "").split(" ");
+
+export interface Spell {
+    Name: string;
+    Id: string;
+    School: string;
+    Classes: string[];
+    Source: string;
+}
+
+export const GetSpellKeywords = (spell: Spell) => _.concat(spell.Classes, spell.School);
+
+export class Library<TItem extends LibraryItem> {
     private items: { [id: string]: TItem } = {};
     private listings: Listing[] = [];
-
+    
     constructor(private route: string, private getKeywords: (item: TItem) => string []) { };
 
     static FromFile<I extends LibraryItem>(filename: string, route: string, getKeywords: (item: I) => string []): Library<I> {
@@ -40,14 +60,14 @@ export default class Library<TItem extends LibraryItem> {
                 throw `Couldn't read ${filename} as a library: ${err}`;
             }
 
-            let newItems: I [] = JSON.parse(buffer.toString());
+            let newItems: any [] = JSON.parse(buffer.toString());
             library.Add(newItems);
         });
 
         return library;
     }
 
-    private Add(items: TItem[]) {
+    private Add(items: any []) {
         items.forEach((c) => {
             c.Id = createId(c.Name, c.Source);
             this.items[c.Id] = c;
@@ -68,4 +88,6 @@ export default class Library<TItem extends LibraryItem> {
     public GetListings(): Listing[] {
         return this.listings;
     }
+
+    public Route = () => this.route;
 }
